renaming algo using a more appropriate name and some refactor
