+++ Cargo.toml
@@ -24,9 +24,5 @@ path = "src/lib.rs"
 [features]
 default = []
 
-[[bench]]
-name = "bench_mandelbrot"
-harness = false
-
 [dev-dependencies]
 criterion = { version = "0.5", features = ["html_reports", "real_blackbox", "cargo_bench_support"] }
