+++ src/lib.rs
@@ -22,8 +22,8 @@ mod ispc_tasks;
 mod scalar_par;
 mod simd_par;
 
-mod compiler_simd_par;
-mod compiler_simd;
+mod autovectorization_par;
+mod autovectorization_ser;
 
 type Range = ops::Range<f64>;
 type Region = (Range, Range);
@@ -38,8 +38,8 @@ pub enum Algorithm {
     Scalar,
     /// Parallel SIMD algorithm using Rayon
     Simd,
-    /// Prallel SIMD generate by compiler
-    CompilerSimd,
+    /// Autovectorization
+    AutoVectorization,
     /// ISPC SIMD + parallel tasks algorithm
     Ispc,
 }
@@ -65,7 +65,7 @@ impl Mandelbrot {
                 scalar_par::generate(dims, region.0, region.1)
             }
             Algorithm::Simd => simd_par::generate(dims, region.0, region.1),
-            Algorithm::CompilerSimd => compiler_simd_par::generate(dims, region.0, region.1),
+            Algorithm::AutoVectorization => autovectorization_par::generate(dims, region.0, region.1),
             #[cfg(feature = "ispc")]
             Algorithm::Ispc => ispc_tasks::generate(dims, region.0, region.1),
             #[cfg(not(feature = "ispc"))]
@@ -74,13 +74,6 @@ impl Mandelbrot {
             ),
         };
 
-        Self { dims, data }
-    }
-
-    pub fn generate_region_cs(
-        dims: Dimensions, region: Region, algo: Algorithm,
-    ) -> Self {
-        let data=compiler_simd::generate(dims, region.0, region.1);
         Self { dims, data }
     }
